+++ test/runtests.jl
@@ -67,7 +67,7 @@ include("common.jl")
 GLIST =    [
             (Graph{Int64}, DiGraph{Int64}),
             (Graph{UInt32}, DiGraph{UInt32}),
-            (Net, DiNet)
+            (Network, DiNetwork)
             ]
 
 for GDG in GLIST, t in tests