+++ NEWS.md
@@ -30,7 +30,7 @@ vertices, edges or the graph itself
 - introduce abstract types `AGraph`, `ADiGraph`  
 - make all methods accept abstract graph types
 - vertex type for Graph can be any integer, i.e. we have Graph{Int32} along the default Graph{Int}
-- add `Net` and `DiNet`, graph types with indexed edges, inspired by graph-tool library
+- add `Network` and `DiNetwork`, graph types with indexed edges, inspired by graph-tool library
 - simplified and generic edge iterator (now a Generator)
 - improve maximum_flow performance
 - add Benchmarks through BenchmarkTools